+++ container_guts/defaults.py
@@ -1,0 +1,6 @@
+__author__ = "Vanessa Sochat"
+__copyright__ = "Copyright 2021-2022, Vanessa Sochat"
+__license__ = "MPL 2.0"
+
+# Default database for base image
+database = "https://github.com/singularityhub/shpc-guts"