+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/singularityhub/guts/tree/main) (0.0.x)
+ - adding fs (filesystem) extraction support (0.0.13)
  - tag should be own file (0.0.12)
  - Support for output directory (so path prepared by guts) (0.0.11)
  - Initial creation of project (0.0.1)